+++ scripts/compile_cost_assumptions.py
@@ -69,12 +69,21 @@ sheet_names = {'onwind': '20 Onshore turbines',
                'biomass HOP': '09c Straw HOP',
                'central coal CHP': '01 Coal CHP',
                'central gas CHP': '04 Gas turb. simple cycle, L',
+               'central gas CHP CC': '04 Gas turb. simple cycle, L',
                'central solid biomass CHP': '09a Wood Chips, Large 50 degree',
+               'central solid biomass CHP CC': '09a Wood Chips, Large 50 degree',
+               'central solid biomass CHP powerboost CC': '09a Wood Chips, Large 50 degree',
+               # 'solid biomass power': '09a Wood Chips extract. plant',
+               # 'solid biomass power CC': '09a Wood Chips extract. plant',
                'central air-sourced heat pump': '40 Comp. hp, airsource 3 MW',
                'central ground-sourced heat pump': '40 Absorption heat pump, DH',
                'central resistive heater': '41 Electric Boilers',
                'central gas boiler': '44 Natural Gas DH Only',
                'decentral gas boiler': '202 Natural gas boiler',
+               'direct firing gas': '312.a Direct firing Natural Gas',
+               'direct firing gas CC': '312.a Direct firing Natural Gas',
+               'direct firing solid fuels': '312.b Direct firing Sold Fuels',
+               'direct firing solid fuels CC': '312.b Direct firing Sold Fuels',
                'decentral ground-sourced heat pump': '207.7 Ground source existing',
                'decentral air-sourced heat pump': '207.3 Air to water existing',
                # 'decentral resistive heater': '216 Electric heating',
@@ -85,6 +94,7 @@ sheet_names = {'onwind': '20 Onshore turbines',
                'hydrogen storage tank type 1 including compressor': '151a Hydrogen Storage - Tanks',
                'micro CHP': '219 LT-PEMFC mCHP - natural gas',
                'biogas' : '81 Biogas Plant, Basic conf.',
+               'biogas CC' : '81 Biogas Plant, Basic conf.',
                'biogas upgrading': '82 Biogas, upgrading',
                'battery': '180 Lithium Ion Battery',
                'industrial heat pump medium temperature': '302.a High temp. hp Up to 125 C',
@@ -92,18 +102,22 @@ sheet_names = {'onwind': '20 Onshore turbines',
                'electric boiler steam': '310.1 Electric boiler steam  ',
                'gas boiler steam': '311.1c Steam boiler Gas',
                'solid biomass boiler steam': '311.1e Steam boiler Wood',
+               'solid biomass boiler steam CC': '311.1e Steam boiler Wood',
                'biomass boiler': '204 Biomass boiler, automatic',
                'electrolysis': '86 AEC 100MW', #'88 Alkaline Electrolyser',
-               'direct air capture' : '403.a Direct air capture',
-               'biomass CHP capture' : '401.a Post comb - small CHP',
-               'cement capture' : '401.c Post comb - Cement kiln',
-               'BioSNG' : '84 Gasif. CFB, Bio-SNG',
-               'BtL' : '85 Gasif. Ent. Flow FT, liq fu ',
+               'direct air capture': '403.a Direct air capture',
+               'biomass CHP capture': '401.a Post comb - small CHP',
+               'cement capture': '401.c Post comb - Cement kiln',
+               'BioSNG': '84 Gasif. CFB, Bio-SNG',
+               'BtL': '85 Gasif. Ent. Flow FT, liq fu ',
                'biogas plus hydrogen': '99 SNG from methan. of biogas',
-               #'methanolisation': '98 Methanol from power',
-               #'Fischer-Tropsch': '102 Hydrogen to Jet',
+               'methanolisation': '98 Methanol from power',
+               'Fischer-Tropsch': '102 Hydrogen to Jet',
+               'central hydrogen CHP': '12 LT-PEMFC CHP',
                'Haber-Bosch': '103 Hydrogen to Ammonia',
                'air separation unit': '103 Hydrogen to Ammonia',
+               'waste CHP': '08 WtE CHP, Large, 50 degree',
+               'waste CHP CC': '08 WtE CHP, Large, 50 degree',
                # 'electricity distribution rural': '101 2 el distri Rural',
                # 'electricity distribution urban': '101 4 el distri  city',
                # 'gas distribution rural': '102 7 gas  Rural',
@@ -129,14 +143,23 @@ uncrtnty_lookup = {'onwind': 'J:K',
                     'biomass HOP': 'I:J',
                     'central coal CHP': '',
                     'central gas CHP': 'I:J',
+                    'central gas CHP CC': 'I:J',
+                    'central hydrogen CHP': 'I:J',
                     'central solid biomass CHP': 'I:J',
-                    # 'central solid biomass CHP CCS': 'I:J',
+                    'central solid biomass CHP CC': 'I:J',
+                    'central solid biomass CHP powerboost CC': 'I:J',
+                    # 'solid biomass power': 'J:K',
+                    # 'solid biomass power CC': 'J:K',
                     'solar': '',
                     'central air-sourced heat pump': 'J:K',
                     'central ground-sourced heat pump': 'I:J',
                     'central resistive heater': 'I:J',
                     'central gas boiler': 'I:J',
                     'decentral gas boiler': 'I:J',
+                    'direct firing gas': 'H:I',
+                    'direct firing gas CC': 'H:I',
+                    'direct firing solid fuels': 'H:I',
+                    'direct firing solid fuels CC': 'H:I',
                     'decentral ground-sourced heat pump': 'I:J',
                     'decentral air-sourced heat pump': 'I:J',
                     'central water tank storage': 'J:K',
@@ -145,6 +168,7 @@ uncrtnty_lookup = {'onwind': 'J:K',
                     'hydrogen storage tank type 1 including compressor': 'J:K',
                     'micro CHP': 'I:J',
                     'biogas': 'I:J',
+                    'biogas CC': 'I:J',
                     'biogas upgrading': 'I:J',
                     'electrolysis': 'I:J',
                     'battery': 'L,N',
@@ -159,11 +183,14 @@ uncrtnty_lookup = {'onwind': 'J:K',
                     'electric boiler steam':'H:I',
                     'gas boiler steam':'H:I',
                     'solid biomass boiler steam':'H:I',
+                    'solid biomass boiler steam CC':'H:I',
                     'biomass boiler': 'I:J',
                     'Fischer-Tropsch': 'I:J',
                     'Haber-Bosch': 'I:J',
                     'air separation unit': 'I:J',
                     'methanolisation': 'J:K',
+                    'waste CHP': 'I:J',
+                    'waste CHP CC': 'I:J',
 }
 
 # since February 2022 DEA uses a new format for the technology data
@@ -218,7 +245,7 @@ def get_data_DEA(tech, data_in, expectation=None):
     elif tech in ['direct air capture', 'cement capture', 'biomass CHP capture']:
         usecols = "A:F"
     elif tech in ['industrial heat pump medium temperature', 'industrial heat pump high temperature',
-                  'electric boiler steam', "gas boiler steam", "solid biomass boiler steam"]:
+                  'electric boiler steam', "gas boiler steam", "solid biomass boiler steam", "solid biomass boiler steam CC", "direct firing gas", "direct firing gas CC", "direct firing solid fuels", "direct firing solid fuels CC"]:
         usecols = "A:E"
     elif tech in ['Fischer-Tropsch', 'Haber-Bosch', 'air separation unit']:
         usecols = "B:F"
@@ -239,6 +266,7 @@ def get_data_DEA(tech, data_in, expectation=None):
                           usecols=usecols,
                           skiprows=skiprows,
                           na_values="N.A")
+    # print(excel)
 
     excel.dropna(axis=1, how="all", inplace=True)
 
@@ -246,6 +274,7 @@ def get_data_DEA(tech, data_in, expectation=None):
     excel.index = excel.index.fillna(" ")
     excel.index = excel.index.astype(str)
     excel.dropna(axis=0, how="all", inplace=True)
+    # print(excel)
 
     if 2020 not in excel.columns:
         selection = excel[excel.isin([2020])].dropna(how="all").index
@@ -286,6 +315,7 @@ def get_data_DEA(tech, data_in, expectation=None):
                   "Output capacity expansion cost",
                   "Hydrogen output",
                   "Hydrogen (% total input_e (MWh / MWh))",
+                  " - hereof recoverable for district heating (%-points of heat loss)",
                   "Cb coefficient",
                   "Cv coefficient",
                   "Distribution network costs", "Technical life",
@@ -294,6 +324,7 @@ def get_data_DEA(tech, data_in, expectation=None):
                   'Heat input', 'Heat  input', 'Electricity input', 'Eletricity input', 'Heat out',
                   'capture rate',
                   "FT Liquids Output, MWh/MWh Total Input",
+                  " - hereof recoverable for district heating (%-points of heat loss)",
                   "Bio SNG (% of fuel input)",
                   "Total O&M"]
 
@@ -314,11 +345,13 @@ def get_data_DEA(tech, data_in, expectation=None):
     # average data  in format "lower_value-upper_value"
     df = df.applymap(lambda x: (float((x).split("-")[0])
                                 + float((x).split("-")[1]))/2 if (type(x)==str and "-" in x) else x)
-    # remove symbols "~", ">", "<"
-    for sym in ["~", ">", "<"]:
+    # remove symbols "~", ">", "<" and " "
+    for sym in ["~", ">", "<", " "]:
         df = df.applymap(lambda x: x.replace(sym,"") if type(x)==str else x)
 
     df = df.astype(float)
+    df = df.mask(df.apply(pd.to_numeric, errors='coerce').isnull(), df.astype(str).apply(lambda x: x.str.strip()))
+    # print(df)
 
     ## Modify data loaded from DEA on a per-technology case
     if (tech == "offwind") and snakemake.config['offwind_no_gridcosts']:
@@ -343,7 +376,7 @@ def get_data_DEA(tech, data_in, expectation=None):
     if tech == 'biomass boiler':
         df.drop(df.loc[df.index.str.contains("Possible additional")].index, inplace=True)
         df.drop(df.loc[df.index.str.contains("Total efficiency")].index, inplace=True)
-	
+
     if tech == "Haber-Bosch":
         df.drop(df.loc[df.index.str.contains("Specific investment mark-up factor optional ASU")].index, inplace=True)
         df.drop(df.loc[df.index.str.contains("Specific investment (MEUR /TPD Ammonia output", regex=False)].index, inplace=True)
@@ -376,6 +409,9 @@ def get_data_DEA(tech, data_in, expectation=None):
         df.drop(df.loc[df.index.str.contains("Fixed O&M (kEUR/MW Ammonia/year)", regex=False)].index, inplace=True)
         df.drop(df.loc[df.index.str.contains("Variable O&M (EUR/MWh Ammonia)", regex=False)].index, inplace=True)
 
+    if "solid biomass power" in tech:
+        df.index = df.index.str.replace("EUR/MWeh", "EUR/MWh")
+
     df_final = pd.DataFrame(index=df.index, columns=years)
 
     # [RTD-interpolation-example]
@@ -399,6 +435,7 @@ def get_data_DEA(tech, data_in, expectation=None):
         df_final["unit"] = (df_final.rename(index=lambda x:
                                             x[x.rfind("(")+1: x.rfind(")")]).index.values)
     df_final.index = df_final.index.str.replace(r" \(.*\)","", regex=True)
+
 
     return df_final
 
@@ -550,7 +587,7 @@ def add_co2_intensity(costs):
     costs.loc[('coal', 'CO2 intensity'), 'value'] = 93369 / 1e3 / TJ_to_MWh  # Steinkohle
     costs.loc[('lignite', 'CO2 intensity'), 'value'] = 113031 / 1e3 / TJ_to_MWh  # Rohbraunkohle Rheinland
     costs.loc[('oil', 'CO2 intensity'), 'value'] = 74020 / 1e3 / TJ_to_MWh  # Heizöl, leicht
-    costs.at[('solid biomass', 'CO2 intensity'), 'value'] = 0.3
+    costs.loc[('solid biomass', 'CO2 intensity'), 'value'] = 0.3
 
     oil_specific_energy = 44 #GJ/t
     CO2_CH2_mass_ratio = 44/14 #kg/kg (1 mol per mol)
@@ -742,6 +779,8 @@ def clean_up_units(tech_data):
     tech_data.unit = tech_data.unit.str.replace("MWh SNG", "MWh_CH4")
     tech_data.unit = tech_data.unit.str.replace("MW SNG", "MW_CH4")
     tech_data.unit = tech_data.unit.str.replace("EUR/MWh of total input", "EUR/MWh_e")
+    tech_data.unit = tech_data.unit.str.replace("EUR/MWeh", "EUR/MWh_e")
+    tech_data.unit = tech_data.unit.str.replace("% -points of heat loss", "MWh_th/MWh_el")
 
 
     tech_data.unit = tech_data.unit.str.replace("FT Liquids Output, MWh/MWh Total Inpu", "MWh_FT/MWh_H2")
@@ -1016,20 +1055,34 @@ def order_data(tech_data):
         efficiency = df[(df.index.str.contains("efficiency") |
                          (df.index.str.contains("Hydrogen output, at LHV"))|
                          (df.index.str.contains("FT Liquids Output, MWh/MWh Total Input"))|
+                         (df.index.str.contains("hereof recoverable for district heating"))|
                          (df.index.str.contains("Bio SNG"))|
                          (df.index == ("Hydrogen")))
                          & ((df.unit=="%") |  (df.unit =="% total size") |
                            (df.unit =="% of fuel input") |
                            (df.unit =="MWh_H2/MWh_e") |
+                           (df.unit =="%-points of heat loss") |
                            df.unit.str.contains("MWh_FT/MWh_H2"))
                          & (~df.index.str.contains("name plate"))].copy()
 
         if tech == 'Fischer-Tropsch':
             efficiency[years] *= 100
 
+
         # take annual average instead of name plate efficiency
         if any(efficiency.index.str.contains("annual average")):
             efficiency = efficiency[efficiency.index.str.contains("annual average")]
+
+        # hydrogen electrolysiswith recoverable heat
+        heat_recovery_label = "hereof recoverable for district heating"
+        with_heat_recovery = efficiency.index.str.contains(heat_recovery_label)
+        if with_heat_recovery.any():
+            efficiency_heat = efficiency[with_heat_recovery].copy()
+            efficiency_heat["parameter"] = "efficiency-heat"
+            clean_df[tech] = pd.concat([clean_df[tech], efficiency_heat])
+            efficiency_h2 = efficiency[efficiency.index.str.contains("Hydrogen")].copy()
+            efficiency_h2["parameter"] = "efficiency"
+            clean_df[tech] = pd.concat([clean_df[tech], efficiency_h2])
 
         # check if electric and heat efficiencies are given
         if (any(["Electric" in ind for ind in efficiency.index]) and
@@ -1230,7 +1283,7 @@ def add_manual_input(data):
 
     # Inflation adjustment for investment and VOM
     mask = df[df['parameter'].isin(['investment','VOM'])].index
-    df.loc[mask, 'value'] /= (1+snakemake.config['rate_inflation'])**(df.loc[mask, 'currency_year']-snakemake.config['eur_year'])
+    df.loc[mask, 'value'] /= (1+snakemake.config['rate_inflation'])**(df.loc[mask, 'currency_year'].astype(int)-snakemake.config['eur_year'])
 
     l = []
     for tech in df['technology'].unique():
@@ -1250,7 +1303,8 @@ def add_manual_input(data):
             l.append(s)
 
     new_df = pd.DataFrame(l).set_index(['technology','parameter'])
-    data = data.combine_first(new_df)
+    # overwrite DEA data with manual input
+    data = new_df.combine_first(data)
 
     return data
 
@@ -1286,18 +1340,34 @@ def carbon_flow(costs,year):
     btlcost_data = np.interp(x=years, xp=[2020, 2050], fp=[3500, 2000])
     btl_cost = pd.Series(data=btlcost_data, index=years)
 
+    bmH2cost_data = np.interp(x=years, xp=[2020, 2050], fp=[4000, 2500])
+    bmH2_cost = pd.Series(data=bmH2cost_data, index=years)
+
     btleta_data = np.interp(x=years, xp=[2020, 2050], fp=[0.35, 0.45])
     btl_eta = pd.Series(data=btleta_data, index=years)
 
+    #Adding pelletizing cost to biomass boiler
+    costs.loc[('biomass boiler', 'pelletizing cost'), 'value'] = 9
+    costs.loc[('biomass boiler', 'pelletizing cost'), 'unit'] = "EUR/MWh_pellets"
+    costs.loc[('biomass boiler', 'pelletizing cost'), 'source'] = "Assumption based on doi:10.1016/j.rser.2019.109506"
 
-    for tech in ['BtL', 'BioSNG', 'methanation', 'Fischer-Tropsch', 'biogas', 'digestible biomass to hydrogen', 'solid biomass to hydrogen']:
+
+    for tech in ['Fischer-Tropsch', 'methanolisation', 'BtL', 'BioSNG', 'biogas',
+                 'biogas CC', 'digestible biomass to hydrogen',
+                 'solid biomass to hydrogen', 'electrobiofuels']:
         inv_cost = 0
         eta = 0
         lifetime = 0
         FOM = 0
         VOM = 0
         source = 'TODO'
-        co2_capture_rate = 0.98
+        co2_capture_rate = 0.90
+
+        if not (tech, "capture rate") in costs.index:
+            costs.loc[(tech, 'capture rate'), 'value'] = co2_capture_rate
+            costs.loc[(tech, 'capture rate'), 'unit'] = "per unit"
+            costs.loc[(tech, 'capture rate'), 'source'] = "Assumption based on doi:10.1016/j.biombioe.2015.01.006"
+
 
         if tech == 'BtL':
             inv_cost = btl_cost[year]
@@ -1305,11 +1375,12 @@ def carbon_flow(costs,year):
             eta = btl_eta[year]
             source = "doi:10.1016/j.enpol.2017.05.013"
 
+
         elif tech == 'BioSNG':
             medium_out = 'gas'
             lifetime = 25
 
-        elif tech == 'biogas':
+        elif tech in ['biogas', 'biogas CC']:
             eta = 1
             source = "Assuming input biomass is already given in biogas output"
             AD_CO2_share = 0.4 #volumetric share in biogas (rest is CH4)
@@ -1335,13 +1406,13 @@ def carbon_flow(costs,year):
             AD_CO2_share = 0.1 #volumetric share in biogas (rest is CH4).
 
         elif tech == 'digestible biomass to hydrogen':
-            inv_cost = 2500
+            inv_cost = bmH2_cost[year]
             eta = 0.39
             FOM = 4.25
             source = 'Zech et.al. DBFZ Report Nr. 19. Hy-NOW - Evaluierung der Verfahren und Technologien für die Bereitstellung von Wasserstoff auf Basis von Biomasse, DBFZ, 2014' #source_dict('HyNOW')
 
         elif tech == 'solid biomass to hydrogen':
-            inv_cost = 2500
+            inv_cost = bmH2_cost[year]
             eta = 0.56
             FOM = 4.25
             source = 'Zech et.al. DBFZ Report Nr. 19. Hy-NOW - Evaluierung der Verfahren und Technologien für die Bereitstellung von Wasserstoff auf Basis von Biomasse, DBFZ, 2014' #source_dict('HyNOW')
@@ -1351,9 +1422,68 @@ def carbon_flow(costs,year):
             costs.loc[(tech, 'efficiency'), 'unit'] = "per unit"
             costs.loc[(tech, 'efficiency'), 'source'] = source
 
-        costs.loc[(tech, 'capture rate'), 'value'] = co2_capture_rate
-        costs.loc[(tech, 'capture rate'), 'unit'] = "per unit"
-        costs.loc[(tech, 'capture rate'), 'source'] = "Assumption based on doi:10.1016/j.biombioe.2015.01.006"
+        if tech in ['BioSNG', 'BtL']:
+            input_CO2_intensity = costs.loc[('solid biomass', 'CO2 intensity'), 'value']
+
+            costs.loc[(tech, 'C in fuel'), 'value'] = costs.loc[(tech, 'efficiency'), 'value'] \
+                                                  * costs.loc[(medium_out, 'CO2 intensity'), 'value'] \
+                                                  / input_CO2_intensity
+            costs.loc[(tech, 'C stored'), 'value'] = 1 - costs.loc[(tech, 'C in fuel'), 'value'] - c_in_char
+            costs.loc[(tech, 'CO2 stored'), 'value'] = input_CO2_intensity * costs.loc[(tech, 'C stored'), 'value']
+
+            costs.loc[(tech, 'C in fuel'), 'unit'] = "per unit"
+            costs.loc[(tech, 'C stored'), 'unit'] = "per unit"
+            costs.loc[(tech, 'CO2 stored'), 'unit'] = "tCO2/MWh_th"
+
+            costs.loc[(tech, 'C in fuel'), 'source'] = "Stoichiometric calculation, doi:10.1016/j.apenergy.2022.120016"
+            costs.loc[(tech, 'C stored'), 'source'] = "Stoichiometric calculation, doi:10.1016/j.apenergy.2022.120016"
+            costs.loc[(tech, 'CO2 stored'), 'source'] = "Stoichiometric calculation, doi:10.1016/j.apenergy.2022.120016"
+
+        elif tech in ['electrobiofuels']:
+
+            input_CO2_intensity = costs.loc[('solid biomass', 'CO2 intensity'), 'value']
+            oil_CO2_intensity = costs.loc[('oil', 'CO2 intensity'), 'value']
+
+            costs.loc[('electrobiofuels', 'C in fuel'), 'value'] = (costs.loc[('BtL', 'C in fuel'), 'value']
+                                                                    + costs.loc[('BtL', 'C stored'), 'value']
+                                                                    * costs.loc[('Fischer-Tropsch', 'capture rate'), 'value'])
+            costs.loc[('electrobiofuels', 'C in fuel'), 'unit'] = 'per unit'
+            costs.loc[('electrobiofuels', 'C in fuel'), 'source'] = 'Stoichiometric calculation'
+
+            costs.loc[('electrobiofuels', 'efficiency-biomass'), 'value'] = costs.loc[('electrobiofuels', 'C in fuel'), 'value'] \
+                                                                            * input_CO2_intensity / oil_CO2_intensity
+            costs.loc[('electrobiofuels', 'efficiency-biomass'), 'unit'] = 'per unit'
+            costs.loc[('electrobiofuels', 'efficiency-biomass'), 'source'] = 'Stoichiometric calculation'
+
+
+            efuel_scale_factor = costs.loc[('BtL', 'C stored'), 'value'] * costs.loc[('Fischer-Tropsch', 'capture rate'), 'value']
+
+            costs.loc[('electrobiofuels', 'efficiency-hydrogen'), 'value'] = costs.loc[('Fischer-Tropsch', 'efficiency'), 'value']\
+                                                                             / efuel_scale_factor
+            costs.loc[('electrobiofuels', 'efficiency-hydrogen'), 'unit'] = 'per unit'
+            costs.loc[('electrobiofuels', 'efficiency-hydrogen'), 'source'] = 'Stoichiometric calculation'
+
+            costs.loc[('electrobiofuels', 'efficiency-tot'), 'value'] = (1 /
+                                                                         (1 / costs.loc[('electrobiofuels', 'efficiency-hydrogen'), 'value'] +
+                                                                          1 / costs.loc[('electrobiofuels', 'efficiency-biomass'), 'value']))
+            costs.loc[('electrobiofuels', 'efficiency-tot'), 'unit'] = 'per unit'
+            costs.loc[('electrobiofuels', 'efficiency-tot'), 'source'] = 'Stoichiometric calculation'
+
+            inv_cost = btl_cost[year] + costs.loc[('Fischer-Tropsch', 'investment'), 'value'] * efuel_scale_factor
+            VOM = costs.loc[('BtL', 'VOM'), 'value'] + costs.loc[('Fischer-Tropsch', 'VOM'), 'value'] * efuel_scale_factor
+            FOM = costs.loc[('BtL', 'FOM'), 'value']
+            medium_out = 'oil'
+            source = "combination of BtL and electrofuels"
+
+        elif tech in ['biogas', 'biogas CC', 'biogas plus hydrogen']:
+            CH4_density = 0.657 #kg/Nm3
+            CO2_density = 1.98 #kg/Nm3
+            CH4_vol_energy_density = CH4_specific_energy * CH4_density / (1000 * 3.6) #MJ/Nm3 -> MWh/Nm3
+            CO2_weight_share = AD_CO2_share * CO2_density
+
+            costs.loc[(tech, 'CO2 stored'), 'value'] = CO2_weight_share / CH4_vol_energy_density / 1000 #tCO2/MWh,in (NB: assuming the input is already given in the biogas potential and cost
+            costs.loc[(tech, 'CO2 stored'), 'unit'] = "tCO2/MWh_th"
+            costs.loc[(tech, 'CO2 stored'), 'source'] = "Stoichiometric calculation, doi:10.1016/j.apenergy.2022.120016"
 
         if inv_cost > 0:
             costs.loc[(tech, 'investment'), 'value'] = inv_cost
@@ -1375,32 +1505,69 @@ def carbon_flow(costs,year):
             costs.loc[(tech, 'VOM'), 'unit'] = "EUR/MWh_th"
             costs.loc[(tech, 'VOM'), 'source'] = source
 
-        if tech in ['BioSNG', 'BtL']:
-            input_CO2_intensity = costs.loc[('solid biomass', 'CO2 intensity'), 'value']
+    return costs
 
-            costs.loc[(tech, 'C in fuel'), 'value'] = costs.loc[(tech, 'efficiency'), 'value'] \
-                                                  * costs.loc[(medium_out, 'CO2 intensity'), 'value'] \
-                                                  / input_CO2_intensity
-            costs.loc[(tech, 'C stored'), 'value'] = 1 - costs.loc[(tech, 'C in fuel'), 'value'] - c_in_char
-            costs.loc[(tech, 'CO2 stored'), 'value'] = input_CO2_intensity * costs.loc[(tech, 'C stored'), 'value']
+def energy_penalty(costs):
 
-            costs.loc[(tech, 'C in fuel'), 'unit'] = "per unit"
-            costs.loc[(tech, 'C stored'), 'unit'] = "per unit"
-            costs.loc[(tech, 'CO2 stored'), 'unit'] = "tCO2/MWh_th"
+    # Energy penalty for biomass carbon capture
+    # Need to take steam production for CC into account, assumed with the main feedstock,
+    # e.g. the input biomass is used also for steam, and the efficiency for el and heat is scaled down accordingly
 
-            costs.loc[(tech, 'C in fuel'), 'source'] = "Stoichiometric calculation"
-            costs.loc[(tech, 'C stored'), 'source'] = "Stoichiometric calculation"
-            costs.loc[(tech, 'CO2 stored'), 'source'] = "Stoichiometric calculation"
+    for tech in ['central solid biomass CHP CC', 'waste CHP CC', 'solid biomass boiler steam CC', 'direct firing solid fuels CC', 'direct firing gas CC', 'biogas CC']:
 
-        elif tech in ['biogas','biogas plus hydrogen']:
-            CH4_density = 0.657 #kg/Nm3
-            CO2_density = 1.98 #kg/Nm3
-            CH4_vol_energy_density = CH4_specific_energy * CH4_density / (1000 * 3.6) #MJ/Nm3 -> MWh/Nm3
-            CO2_weight_share = AD_CO2_share * CO2_density
+        if 'powerboost' in tech:
+            boiler = 'electric boiler steam'
+            feedstock = 'solid biomass'
+            co2_capture = costs.loc[(feedstock, 'CO2 intensity'), 'value']
+        elif 'gas' in tech:
+            boiler = 'gas boiler steam'
+            feedstock = 'gas'
+            co2_capture = costs.loc[(feedstock, 'CO2 intensity'), 'value']
+        elif 'biogas' in tech:
+            boiler = 'gas boiler steam'
+            co2_capture = costs.loc[(tech, 'CO2 stored'), 'value']
+        else:
+            boiler = 'solid biomass boiler steam'
+            feedstock = 'solid biomass'
+            co2_capture = costs.loc[(feedstock, 'CO2 intensity'), 'value']
 
-            costs.loc[(tech, 'CO2 stored'), 'value'] = CO2_weight_share / CH4_vol_energy_density / 1000 #tCO2/MWh,in (NB: assuming the input is already given in the biogas potential and cost
-            costs.loc[(tech, 'CO2 stored'), 'unit'] = "tCO2/MWh_th"
-            costs.loc[(tech, 'CO2 stored'), 'source'] = "Stoichiometric calculation"
+        #Scaling biomass input to account for heat demand of carbon capture
+        scalingFactor = 1 / (1 + co2_capture * costs.loc[
+            ('biomass CHP capture', 'heat-input'), 'value']
+                             / costs.loc[(boiler, 'efficiency'), 'value'])
+
+        el_demand = (co2_capture * costs.loc[('biomass CHP capture', 'heat-input'), 'value']
+                     / costs.loc[(boiler, 'efficiency'), 'value'])
+        eta_steam = (1 - scalingFactor) * costs.loc[(boiler, 'efficiency'), 'value']
+        eta_old = costs.loc[(tech, 'efficiency'), 'value']
+
+        # Adapting investment share of tech due to steam boiler addition. Investment per MW_el.
+        costs.loc[(tech, 'investment'), 'value'] = costs.loc[(tech, 'investment'), 'value'] * eta_old / eta_main \
+            + costs.loc[(boiler, 'investment'), 'value'] * eta_steam / eta_main
+        costs.loc[(tech, 'investment'), 'source'] = 'Combination of ' + tech + ' and ' + boiler
+        costs.loc[(tech, 'investment'), 'further description'] = ''
+
+        costs.loc[(tech, 'efficiency'), 'value'] = eta_main
+        costs.loc[(tech, 'efficiency'), 'source'] = 'Combination of ' + tech + ' and ' + boiler
+        costs.loc[(tech, 'efficiency'), 'further description'] = ''
+
+        if 'CHP' in tech:
+            costs.loc[(tech, 'efficiency-heat'), 'value'] = \
+                costs.loc[(tech, 'efficiency-heat'), 'value'] * scalingFactor \
+                    + costs.loc[('solid biomass', 'CO2 intensity'), 'value'] * \
+                (costs.loc[('biomass CHP capture', 'heat-output'), 'value'] +
+                 costs.loc[('biomass CHP capture', 'compression-heat-output'), 'value'])
+            costs.loc[(tech, 'efficiency-heat'), 'source'] = 'Combination of ' + tech + ' and ' + boiler
+            costs.loc[(tech, 'efficiency-heat'), 'further description'] = ''
+
+        if 'biogas CC' in tech:
+            costs.loc[(tech, 'VOM'), 'value'] = 0
+            costs.loc[(tech, 'VOM'), 'unit'] = 'EUR/MWh'
+
+        costs.loc[(tech, 'VOM'), 'value'] = costs.loc[(tech, 'VOM'), 'value'] * eta_old / eta_main \
+            + costs.loc[(boiler, 'VOM'), 'value'] * eta_steam / eta_main
+        costs.loc[(tech, 'VOM'), 'source'] = 'Combination of ' + tech + ' and ' + boiler
+        costs.loc[(tech, 'VOM'), 'further description'] = ''
 
     return costs
 
@@ -1582,7 +1749,7 @@ if __name__ == "__main__":
         from _helpers import mock_snakemake
         os.chdir(os.path.join(os.getcwd(), "scripts"))
         snakemake = mock_snakemake("compile_cost_assumptions")
-    
+
     years = snakemake.config['years']
 
     # (1) DEA data
@@ -1687,6 +1854,9 @@ if __name__ == "__main__":
 
         #carbon balances
         costs = carbon_flow(costs,year)
+
+        #energy penalty of carbon capture
+        costs = energy_penalty(costs)
 
         # include old pypsa costs
         check = pd.concat([costs_pypsa, costs], sort=True, axis=1)